fix: scope demo analytics
